requirements.txt
alembic==1.7.5
boto3==1.20.10
botocore==1.23.10
click==8.0.3
Faker==9.8.0
Flask==2.0.2
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
pathlib==1.0.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.26
text-unidecode==1.3
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
